requirements.txt
# Must
django
djangorestframework